Added exception and completer Message service logic.

+++ src/main/java/com/example/messaging_api/service/MessageService.java
@@ -1,10 +1,16 @@
 package com.example.messaging_api.service;
 
+import com.example.messaging_api.exceptions.InformationNotFoundException;
 import com.example.messaging_api.model.Message;
+import com.example.messaging_api.model.User;
 import com.example.messaging_api.repository.MessageRepository;
+import com.example.messaging_api.repository.UserRepository;
+import com.example.messaging_api.security.MyUserDetails;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
+import java.time.LocalDate;
 import java.util.List;
 import java.util.Optional;
 
@@ -14,24 +20,36 @@ public class MessageService {
     private MessageRepository messageRepository;
 
     @Autowired
-
     public void setMessageRepository(MessageRepository messageRepository) {
         this.messageRepository = messageRepository;
     }
 
+    private UserRepository userRepository;
+
+    @Autowired
+    public void setUserRepository(UserRepository userRepository) {
+        this.userRepository = userRepository;
+    }
+
     public List<Message> getAllMessages(){
-        List<Message> messages = messageRepository.findAll();
+        MyUserDetails userDetails = (MyUserDetails) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        List<Message> messages = messageRepository.findBySenderIdOrReceiverId(userDetails.getUser().getId(), userDetails.getUser().getId());
         return messages;
     }
 
     public Message createMessageByReceiverId(Long receiverId, Message messageObject) {
-//        User receiver = userRepository.findById(receiverId);
-//        messageObject.setReceiver(receiver);
+        MyUserDetails userDetails = (MyUserDetails) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        messageObject.setSender(userDetails.getUser());
+        Optional<User> receiver = userRepository.findById(receiverId);
+        messageObject.setReceiver(receiver.get());
+        messageObject.setTimestamp(LocalDate.now());
+        messageObject.setRead(false);
         return messageRepository.save(messageObject);
     }
 
     public List<Message> getMessagesByReceiverId(Long receiverId) {
-        List<Message> messages = messageRepository.findByReceiverId(receiverId);
+        MyUserDetails userDetails = (MyUserDetails) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        List<Message> messages = messageRepository.findBySenderIdAndReceiverId(userDetails.getUser().getId(), receiverId);
         return messages;
     }
 
@@ -42,8 +60,14 @@ public class MessageService {
     }
 
     public Message deleteMessage(Long messageId) {
-        Optional<Message> message = messageRepository.findById(messageId);
-        messageRepository.deleteById(messageId);
-        return message.get();
+        MyUserDetails userDetails = (MyUserDetails) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        Optional<Message> message = messageRepository.findByIdAndSenderId(messageId, userDetails.getUser().getId());
+        if(message.isEmpty()){
+            throw new InformationNotFoundException("User with id " + userDetails.getUser().getId() + " does not have message with Id " + messageId);
+        } else {
+            messageRepository.deleteById(messageId);
+            return message.get();
+        }
+
     }
 }
